[REFACTOR] 인증 컨트롤러 구조 개선 - 비즈니스 로직 분리 및 응답 표준화
* AuthController 리팩토링 - ApiController 상속 및 ResponseUtils 적용
* AuthService 신규 생성 - 인증 관련 비즈니스 로직 분리
* 로그인/토큰갱신/로그아웃 API 응답 구조 표준화
* LoginResponseDto, TokenRefreshResponseDto 적용으로 타입 안전성 확보
* Controller 책임 분리 - 쿠키 처리는 Controller, 인증 로직은 Service
* 요청/응답에 필요한 LoginResponseDto와 TokenRefreshResponseDto 생성
* AuthException 커스텀 예외 적용 - 명확한 인증 오류 처리
* 상세 로깅 추가 - 로그인/로그아웃/토큰갱신 이벤트 기록
* JavaDoc 주석 추가로 메서드 역할 명확화
