+++ src/main/java/hashsnap/login/dto/LoginResponseDto.java
@@ -1,0 +1,12 @@
+package hashsnap.login.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class LoginResponseDto {
+    private String accessToken;
+    private String refreshToken;
+    private String userEmail;
+}
