+++ src/main/java/hashsnap/login/dto/TokenRefreshResponseDto.java
@@ -1,0 +1,10 @@
+package hashsnap.login.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class TokenRefreshResponseDto {
+    private String accessToken;
+}
